// src/app/api/model/collectionDetails.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CollectionPart } from './collectionPart';


export interface CollectionDetails { 
    id?: number;
    name?: string;
    original_language?: string;
    original_name?: string;
    overview?: string;
    poster_path?: string | null;
    backdrop_path?: string | null;
    parts?: Array<CollectionPart>;
}